Move JiraTicketList to common location
